refactor: address PR review comments
See https://github.com/ibi-group/datatools-server/pull/381
