+++ src/main/java/com/conveyal/datatools/manager/controllers/api/DeploymentController.java
@@ -18,7 +18,6 @@ import com.conveyal.datatools.manager.models.Project;
 import com.conveyal.datatools.manager.persistence.Persistence;
 import com.conveyal.datatools.manager.utils.JobUtils;
 import com.conveyal.datatools.manager.utils.json.JsonManager;
-import com.mongodb.client.FindIterable;
 import org.bson.Document;
 import org.eclipse.jetty.http.HttpStatus;
 import org.slf4j.Logger;
@@ -32,9 +31,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 import static com.conveyal.datatools.common.utils.SparkUtils.logMessageAndHalt;
@@ -50,7 +47,6 @@ import static spark.Spark.put;
  */
 public class DeploymentController {
     private static final Logger LOG = LoggerFactory.getLogger(DeploymentController.class);
-    private static final Map<String, DeployJob> deploymentJobsByServer = new HashMap<>();
 
     /**
      * Gets the deployment specified by the request's id parameter and ensure that user has access to the
@@ -378,45 +374,6 @@ public class DeploymentController {
     }
 
     /**
-     * Creates and queues a new {@link DeployJob} if there are no conflicting jobs assigned to the specified server.
-     *
-     * @param deployment The deployment to associate the new DeployJob with
-     * @param owner The owner to associate the new DeployJob with
-     * @param server The server to associate the new DeployJob with
-     * @return returns the DeployJob if the job was successfully queued, otherwise this returns null
-     */
-    public static DeployJob queueDeployJob(Deployment deployment, Auth0UserProfile owner, OtpServer server) {
-        // Check that we can deploy to the specified target. (Any deploy job for the target that is presently active will
-        // cause a halt.)
-        if (deploymentJobsByServer.containsKey(server.id)) {
-            // There is a deploy job for the server. Check if it is active.
-            DeployJob conflictingDeployJob = deploymentJobsByServer.get(server.id);
-            if (conflictingDeployJob != null && !conflictingDeployJob.status.completed) {
-                // Another deploy job is actively being deployed to the server target.
-                LOG.error("New deploy job will not be queued due to active deploy job in progress.");
-                return null;
-            }
-        }
-
-        // For any previous deployments sent to the server/router combination, set deployedTo to null because
-        // this new one will overwrite it. NOTE: deployedTo for the current deployment will only be updated after the
-        // successful completion of the deploy job.
-        FindIterable<Deployment> deploymentsWithSameTarget = Deployment.retrieveDeploymentForServerAndRouterId(
-            server.id,
-            deployment.routerId
-        );
-        for (Deployment oldDeployment : deploymentsWithSameTarget) {
-            LOG.info("Setting deployment target to null for id={}", oldDeployment.id);
-            Persistence.deployments.updateField(oldDeployment.id, "deployedTo", null);
-        }
-        // Finally, add deploy job to the heavy executor.
-        DeployJob deployJob = new DeployJob(deployment, owner, server);
-        JobUtils.heavyExecutor.execute(deployJob);
-        deploymentJobsByServer.put(server.id, deployJob);
-        return deployJob;
-    }
-
-    /**
      * HTTP controller to fetch information about provided EC2 machines that power ELBs running a trip planner.
      */
     private static List<EC2InstanceSummary> fetchEC2InstanceSummaries(
@@ -463,7 +420,7 @@ public class DeploymentController {
         }
 
         // Execute the deployment job and keep track of it in the jobs for server map.
-        DeployJob job = queueDeployJob(deployment, userProfile, otpServer);
+        DeployJob job = JobUtils.queueDeployJob(deployment, userProfile, otpServer);
         if (job == null) {
             // Job for the target is still active! Send a 202 to the requester to indicate that it is not possible
             // to deploy to this target right now because someone else is deploying.
